Break RowSelection logic into topical files for easier reading.
