+++ src/contexts/RowSelection/RowSelectionTypes.ts
@@ -1,0 +1,6 @@
+export type SortingRule = {
+    columnName: string,
+    sortAscending: boolean
+}
+
+export type SortingCallback = (rules: SortingRule[]) => number[]
