+++ src/views/common/SortableTableWidget/SortableTableWidgetHeader.tsx
@@ -1,11 +1,12 @@
-import { faCaretDown, faCaretUp } from '@fortawesome/free-solid-svg-icons'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { Grid, LinearProgress, TableCell, TableHead, TableRow } from '@material-ui/core'
-import { RowSelectionAction, RowSelectionState, SortingRule, TOGGLE_SELECT_ALL, UPDATE_SORT_FIELDS } from 'contexts/RowSelectionContext'
-import React, { FunctionComponent, useCallback, useMemo } from 'react'
-import './SortableTableWidget.css'
-import SortableTableWidgetCheckbox from './SortableTableWidgetCheckbox'
-import { ColsDict, RowsDict, SortableTableWidgetColumn } from './SortableTableWidgetTypes'
+import { faCaretDown, faCaretUp } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { Grid, LinearProgress, TableCell, TableHead, TableRow } from '@material-ui/core';
+import { RowSelectionAction, RowSelectionState, TOGGLE_SELECT_ALL, UPDATE_SORT_FIELDS } from 'contexts/RowSelection/RowSelectionContext';
+import { SortingRule } from 'contexts/RowSelection/RowSelectionTypes';
+import React, { FunctionComponent, useCallback, useMemo } from 'react';
+import './SortableTableWidget.css';
+import SortableTableWidgetCheckbox from './SortableTableWidgetCheckbox';
+import { ColsDict, RowsDict, SortableTableWidgetColumn } from './SortableTableWidgetTypes';
 
 type HeaderRowProps = {
     columns: SortableTableWidgetColumn[]
